+++ yellowbox_heksher/heksher_service.py
@@ -5,25 +5,25 @@ import httpx
 import requests  # type: ignore
 from docker import DockerClient
 from requests import HTTPError, exceptions
-from yellowbox.containers import SafeContainerCreator, get_ports
+from yellowbox.containers import create_and_pull, get_ports
 from yellowbox.extras.postgresql import PostgreSQLService
 from yellowbox.networks import anonymous_network
 from yellowbox.retry import RetrySpec
-from yellowbox.subclasses import RunMixin, SingleEndpointService
+from yellowbox.subclasses import RunMixin, SingleContainerService
 from yellowbox.utils import docker_host_name
 
 
-class HeksherService(SingleEndpointService, RunMixin):
+class HeksherService(SingleContainerService, RunMixin):
     def __init__(self, docker_client: DockerClient, postgres_image: str = 'postgres:latest',
                  heksher_image: str = 'biocatchltd/heksher:0.4.1', port: int = 0, *,
                  heksher_startup_context_features: str, **kwargs):
         self.heksher_image = heksher_image
-        self.container_creator = SafeContainerCreator(docker_client)
+        self.docker_client = docker_client
 
         self.postgres_service = PostgreSQLService(docker_client, image=postgres_image, default_db='heksher')
 
-        self.heksher = self.container_creator.create_and_pull(
-            heksher_image, publish_all_ports=True, ports={80: port}, detach=True, environment={
+        self.heksher = create_and_pull(
+            docker_client, heksher_image, publish_all_ports=True, ports={80: port}, detach=True, environment={
                 'HEKSHER_DB_CONNECTION_STRING': self.postgres_service.container_connection_string("postgres"),
                 'HEKSHER_STARTUP_CONTEXT_FEATURES': heksher_startup_context_features,
             }
@@ -34,7 +34,7 @@ class HeksherService(SingleEndpointService, RunMixin):
         self.network = anonymous_network(docker_client)
         self.postgres_service.connect(self.network, aliases=['postgres'])
         self.network.connect(self.heksher, aliases=['heksher'])
-        super().__init__((self.heksher, ), **kwargs)
+        super().__init__(self.heksher, **kwargs)
 
     def start(self, retry_spec: Optional[RetrySpec] = None):
         self.postgres_service.start()
@@ -106,13 +106,14 @@ class HeksherService(SingleEndpointService, RunMixin):
             (await self.http_client.delete(f'/api/v1/settings/{setting_name}')).raise_for_status()
 
     def _run_db_migration(self):
-        heksher_db_migrator = self.container_creator.create_and_pull(
-            self.heksher_image, "alembic upgrade head", publish_all_ports=True, detach=True, environment={
+        heksher_db_migrator = create_and_pull(
+            self.docker_client, self.heksher_image, "alembic upgrade head", publish_all_ports=True, detach=True,
+            environment={
                 'HEKSHER_DB_CONNECTION_STRING': self.postgres_service.container_connection_string("postgres"),
             }
         )
         self.network.connect(heksher_db_migrator, aliases=['heksher-db-migrator'])
         heksher_db_migrator.start()
-        heksher_db_migrator.stop()
+        heksher_db_migrator.wait()
         self.network.disconnect(heksher_db_migrator)
         heksher_db_migrator.remove(v=True)